rabbitmq added in design and order

+++ order-service/src/main/java/com/stackroute/orderservice/dto/CreateOrderDTO.java
@@ -32,5 +32,5 @@ public class CreateOrderDTO {
     @NotBlank(message = "Design id can't be blank or empty")
     String designId;
     String designName;
-    Double designPrice;
+    Integer designPrice;
 }
